adjust utility class
